WIP - refactoring init generator and adding test for questioner
