+++ package.json
@@ -66,6 +66,7 @@
     "handlebars": "^4.0.11",
     "handlebars-loader": "^1.7.0",
     "http-server": "^0.9.0",
+    "inquirer-test": "^1.0.5",
     "less": "^2.7.2",
     "less-loader": "^4.0.2",
     "mocha": "^3.2.0",
@@ -78,6 +79,7 @@
     "sinon": "^2.3.4",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.1",
+    "tmp": "^0.0.33",
     "ts-loader": "^2.1.0",
     "typescript": "^2.3.4",
     "url-loader": "^1.0.1",
